+++ ndk/src/main/jni/report.h
@@ -315,7 +315,9 @@ void bugsnag_report_set_user_email(bugsnag_report *report, char *value);
 void bugsnag_report_set_user_id(bugsnag_report *report, char *value);
 void bugsnag_report_set_user_name(bugsnag_report *report, char *value);
 void bugsnag_report_start_session(bugsnag_report *report, char *session_id,
-                                  char *started_at);
+                                  char *started_at, int handled_count);
+bool bugsnag_report_has_session(bugsnag_report *report);
+
 #ifdef __cplusplus
 }
 #endif